refactor(integrations): catalog page walk lives in composio-catalog.ts
